Reorganize under leaderboard subpackage
